OMG! The Yelp API!  And Filtering ....DONE!
The implementation of Yelp took me forever just to realize it requires
server, and will not work with CORS or JSONP workarounds.  The
filtering was not too bad, but had to learn how to control variables on
so many layers within function within functions…It now works…My project
is now complete.